Annotations are updated to add more precision to the hexagonal documentation
