+++ domain/src/main/kotlin/fr/sacane/jmanager/domain/hexadoc/DomainAnnotation.kt
@@ -1,18 +1,18 @@
 package fr.sacane.jmanager.domain.hexadoc
 
-
 @Retention(AnnotationRetention.SOURCE)
 @Target(AnnotationTarget.CLASS)
-annotation class LeftPort
-
-@Retention(AnnotationRetention.SOURCE)
-@Target(AnnotationTarget.CLASS)
-annotation class RightPort
-
-@Retention(AnnotationRetention.SOURCE)
-@Target(AnnotationTarget.CLASS)
-annotation class DomainEntity
+annotation class Port(val domainSide: DomainSide = DomainSide.DATASOURCE)
 
 @Retention(AnnotationRetention.SOURCE)
 @Target(AnnotationTarget.CLASS)
 annotation class DefaultSource
+
+@Retention(AnnotationRetention.SOURCE)
+@Target(AnnotationTarget.CLASS)
+annotation class DomainImplementation
+
+enum class DomainSide{
+    DATASOURCE,
+    API
+}
