+++ domain/src/main/kotlin/fr/sacane/jmanager/domain/hexadoc/LeftAnnotation.kt
@@ -4,7 +4,3 @@ package fr.sacane.jmanager.domain.hexadoc
 @Retention(AnnotationRetention.SOURCE)
 @Target(AnnotationTarget.CLASS)
 annotation class LeftAdapter
-
-@Retention(AnnotationRetention.SOURCE)
-@Target(AnnotationTarget.CLASS)
-annotation class LeftEndpoint
