+++ domain/src/main/kotlin/fr/sacane/jmanager/domain/hexadoc/RightAnnotation.kt
@@ -3,7 +3,3 @@ package fr.sacane.jmanager.domain.hexadoc
 @Retention(AnnotationRetention.SOURCE)
 @Target(AnnotationTarget.CLASS)
 annotation class DatasourceAdapter
-
-@Retention(AnnotationRetention.SOURCE)
-@Target(AnnotationTarget.CLASS)
-annotation class DatasourceEntity
