+++ domain/src/main/kotlin/fr/sacane/jmanager/domain/port/spi/LoginManager.kt
@@ -1,9 +1,11 @@
 package fr.sacane.jmanager.domain.port.spi
 
+import fr.sacane.jmanager.domain.hexadoc.DomainSide
+import fr.sacane.jmanager.domain.hexadoc.Port
 import fr.sacane.jmanager.domain.hexadoc.RightPort
 import fr.sacane.jmanager.domain.models.*
 
-@RightPort
+@Port(DomainSide.DATASOURCE)
 interface LoginManager {
     fun login(userPseudonym: String, password: Password): Ticket?
     fun logout(userId: UserId, token: Token): Token?
