+++ domain/src/main/kotlin/fr/sacane/jmanager/domain/port/spi/TransactionRegister.kt
@@ -1,8 +1,10 @@
 package fr.sacane.jmanager.domain.port.spi
+import fr.sacane.jmanager.domain.hexadoc.DomainSide
+import fr.sacane.jmanager.domain.hexadoc.Port
 import fr.sacane.jmanager.domain.hexadoc.RightPort
 import fr.sacane.jmanager.domain.models.*
 
-@RightPort
+@Port(DomainSide.DATASOURCE)
 interface TransactionRegister {
 
     companion object{
