+++ domain/src/main/kotlin/fr/sacane/jmanager/domain/port/spi/UserDirectory.kt
@@ -1,9 +1,11 @@
 package fr.sacane.jmanager.domain.port.spi
 
-import fr.sacane.jmanager.domain.hexadoc.RightPort
+import fr.sacane.jmanager.domain.hexadoc.DomainSide
+import fr.sacane.jmanager.domain.hexadoc.Port
 import fr.sacane.jmanager.domain.models.*
 
-@RightPort
+
+@Port(DomainSide.DATASOURCE)
 interface UserTransaction {
     fun findById(userId: UserId): Ticket?
     fun findUserById(userId: UserId): User?
