+++ infra/src/main/kotlin/fr/sacane/jmanager/infrastructure/server/entity/CategoryResource.kt
@@ -1,10 +1,11 @@
 package fr.sacane.jmanager.infrastructure.server.entity
 
-import fr.sacane.jmanager.domain.hexadoc.DatasourceEntity
-import jakarta.persistence.*
+import jakarta.persistence.Column
+import jakarta.persistence.Entity
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.Id
 
 @Entity
-@DatasourceEntity
 class CategoryResource(
     @Id
     @GeneratedValue
