+++ infra/src/main/kotlin/fr/sacane/jmanager/infrastructure/server/entity/UserResource.kt
@@ -1,13 +1,11 @@
 package fr.sacane.jmanager.infrastructure.server.entity
 
-import fr.sacane.jmanager.domain.hexadoc.DatasourceEntity
-import org.springframework.lang.Nullable
 import jakarta.persistence.*
+import org.springframework.lang.Nullable
 
 
 @Table(name="userResource")
 @Entity
-@DatasourceEntity
 class UserResource(
     @Id
     @Nullable
